Move Moment.js wrapper in the Utils directory instead  of actions
The main motivation is that the wrapper is a utils class and not an action.
