+++ app/utils/actions/CreateNewPersonalTimeZone.js
@@ -12,7 +12,7 @@
  * You should have received a copy of the GNU General Public License along with this program;
  * if not, write to the Free Software Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
-import MomentTimeZone from './MomentTimeZone';
+import MomentTimeZone from '../MomentTimeZone';
 import TimeZoneCity from '../model/TimeZoneCity';
 import Util from '../Util';
